+++ src/utils/domTarget.ts
@@ -1,4 +1,4 @@
-import { Ref } from "vue";
+import type { ComputedRef, Ref } from "vue";
 import { isFunction } from "./index";
 import isBrowser from "./isBrowser";
 
@@ -9,7 +9,8 @@ type TargetType = HTMLElement | Element | Window | Document;
 export type BasicTarget<T extends TargetType = Element> =
   | (() => TargetValue<T>)
   | TargetValue<T>
-  | Ref<TargetValue<T>>;
+  | Ref<TargetValue<T>>
+  | ComputedRef<TargetValue<T>>;
 
 export function getTargetElement<T extends TargetType>(
   target: BasicTarget<T>,