+++ src/utils/useEffectWithTarget.ts
@@ -1,0 +1,5 @@
+import createEffectWithTarget from "./createEffectWithTarget";
+
+const useEffectWithTarget = createEffectWithTarget("useEffect");
+
+export default useEffectWithTarget;